Fixes to homework 3 answers.
Problem 3(a): diagram reversed A and B.

Problem 6(a): reversed \cup and \cap in some places.

Problem 6(b): it correctly states that A and C are dependent, but
the argument is incorrect.  Fixed.

Added note about how P(A\cap B) means the same thing as P(AB).
Use whichever you like.