Minor fix to typo.
Problem 2(c), I used $\nabla_x f(\mathbf{x})$ which leaves $x$
as non-bold, which looks like I am saying $x$ is a scalar.
If $x$ is a vector, it should be represented as $\mathbf{x}$.
Thus the nabla (gradient) should be appear as
$\nabla_{\mathbf{x}} f(\mathbf{x}$ or $\nabla f(\mathbf{x}$.
The latter omits the subscript which is interepted as meaning
all arguments (i.e., the entire vector) is included in the gradient.
I chose to drop the subscript.

I don't think this difference is consequential enough that it probably
tripped anyone up.

I used $\nabla$ in problem 2(b) w